+++ app/Http/Livewire/Backend/RolesTable.php
@@ -3,31 +3,15 @@
 namespace App\Http\Livewire\Backend;
 
 use App\Domains\Auth\Models\Role;
-use App\Domains\Auth\Models\User;
 use Illuminate\Database\Eloquent\Builder;
-use Rappasoft\LaravelLivewireTables\TableComponent;
-use Rappasoft\LaravelLivewireTables\Traits\HtmlComponents;
+use Rappasoft\LaravelLivewireTables\DataTableComponent;
 use Rappasoft\LaravelLivewireTables\Views\Column;
 
 /**
  * Class RolesTable.
  */
-class RolesTable extends TableComponent
+class RolesTable extends DataTableComponent
 {
-    use HtmlComponents;
-
-    /**
-     * @var string
-     */
-    public $sortField = 'name';
-
-    /**
-     * @var array
-     */
-    protected $options = [
-        'bootstrap.container' => false,
-        'bootstrap.classes.table' => 'table table-striped',
-    ];
 
     /**
      * @return Builder
@@ -35,46 +19,26 @@ class RolesTable extends TableComponent
     public function query(): Builder
     {
         return Role::with('permissions:id,name,description')
-            ->withCount('users');
+            ->withCount('users')
+            ->when($this->getFilter('search'), fn ($query, $term) => $query->search($term));
     }
 
-    /**
-     * @return array
-     */
     public function columns(): array
     {
         return [
-            Column::make(__('Type'), 'type')
-                ->sortable()
-                ->format(function (Role $model) {
-                    if ($model->type === User::TYPE_ADMIN) {
-                        return __('Administrator');
-                    }
-
-                    if ($model->type === User::TYPE_USER) {
-                        return __('User');
-                    }
-
-                    return 'N/A';
-                }),
-            Column::make(__('Name'), 'name')
-                ->searchable()
+            Column::make(__('Type'))
                 ->sortable(),
-            Column::make(__('Permissions'), 'permissions_label')
-                ->searchable(function ($builder, $term) {
-                    return $builder->orWhereHas('permissions', function ($query) use ($term) {
-                        return $query->where('name', 'like', '%'.$term.'%');
-                    });
-                })
-                ->format(function (Role $model) {
-                    return $this->html($model->permissions_label);
-                }),
+            Column::make(__('Name'))
+                ->sortable(),
+            Column::make(__('Permissions')),
             Column::make(__('Number of Users'), 'users_count')
                 ->sortable(),
-            Column::make(__('Actions'))
-                ->format(function (Role $model) {
-                    return view('backend.auth.role.includes.actions', ['model' => $model]);
-                }),
+            Column::make(__('Actions')),
         ];
+    }
+
+    public function rowView(): string
+    {
+        return 'backend.auth.role.includes.row';
     }
 }